refactor(react): update radio button group tests
