+++ packages/react/src/components/RadioButtonGroup/index.js
@@ -5,12 +5,6 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import * as FeatureFlags from '@carbon/feature-flags';
-import { default as RadioButtonGroupNext } from './next/RadioButtonGroup';
-import { default as RadioButtonGroupClassic } from './RadioButtonGroup';
-
-const RadioButtonGroup = FeatureFlags.enabled('enable-v11-release')
-  ? RadioButtonGroupNext
-  : RadioButtonGroupClassic;
+import RadioButtonGroup from './RadioButtonGroup';
 
 export default RadioButtonGroup;
